Imports::Uncategorized#process
To process an uncategorized import, we have three steps:

1. Categorize it,
2. Process it according to the category,
3. Mark ourselves as processed.

To categorize an import, we need to know the filename suffix and whether
it's part of a listing (`Scraper::ApprenticeshipBulletinsJob`):

- `.doc` and `.pdf` are straightforward.
- `.docx` is either an Imports::Docx or an Imports::DocxListing,
  depending on whether we came from ApprenticeshipBulletinsJob.

So extend `#process` to take flags that we can pass through as
appropriate. Each import can make use of the flag as it sees fit. This
allows for extension later as we add more ways to acquire imports.

<https://app.asana.com/0/1203289004376659/1207002310256043>